refactor: test의 core 패키지 제거
